Shader SSBO's lighting, still need to fix the distance and per object materials

+++ src/main/java/com/kronos/graphixs/LoopAccessor.java
@@ -4,7 +4,7 @@ public interface LoopAccessor {
 
 	public double getFps();
 
-	public float getDeltaTime();
+	public double getDeltaTime();
 
 	public void setFps(int fps);
